app/requirements.txt
certifi==2020.11.8
chardet==3.0.4
Django==2.2.17
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
Markdown==3.3.3
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
